Sauvegarde avant Armaggedon (extraction t_process)

+++ srcs/VM/functions_opcode/opcode_sti.c
@@ -43,7 +43,7 @@ int				opcode_sti(t_var *data, t_process *p_process)
 
 	if (!check_sti(data, p_process->pc) && !ft_params_opcode(data, p_process, 2, 0))
 	{
-		tmp_adr = ((p_process->pc + data->t_params[0][1] + data->t_params[0][2]) % MEM_SIZE);
+		tmp_adr = ((p_process->pc + (short)data->t_params[0][1] + (short)data->t_params[0][2]) % MEM_SIZE);
 		tmp_val[0] = (unsigned char)((data->t_params[0][0]) >> 24);
 		tmp_val[1] = (unsigned char)((data->t_params[0][0]) >> 16);
 		tmp_val[2] = (unsigned char)((data->t_params[0][0]) >> 8);
@@ -57,11 +57,11 @@ int				opcode_sti(t_var *data, t_process *p_process)
 		if (data->v == 4 || data->v == 6)
 		{
 			ft_printf("P %4i | sti r%i %i %i\n", p_process->id,
-					  data->t_params[1][0], data->t_params[0][1],
+					  (short)data->t_params[1][0], (short)data->t_params[0][1],
 					  data->t_params[0][2]);
 			ft_printf("       | -> store to %i + %i = %i (with pc and mod %i)\n",
-					  data->t_params[0][1], data->t_params[0][2],
-					  (data->t_params[0][1] + data->t_params[0][2]), tmp_adr);
+					  (short)data->t_params[0][1], (short)data->t_params[0][2],
+					  ((short)data->t_params[0][1] + (short)data->t_params[0][2]), tmp_adr);
 		}
 		p_process->pc =  ((p_process->pc + data->op_size) % MEM_SIZE);
 		return (EXIT_SUCCESS);
